+++ src/app/components/plan/PlanStatisticsTables.jsx
@@ -4,7 +4,6 @@ import {
     deltaColor,
     deltaPct,
     fmtDate,
-    fmtFloat,
     fmtInt,
     fmtMinutesAsHHMM,
     fmtPct